Diploma2/pyproject.toml
[tool.poetry]
name = "diploma"
version = "0.1.0"
description = ""
authors = ["RafPC <g.raf1279@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
update = "^0.0.1"
social-auth-app-django = "^5.2.0"
install = "^1.3.5"
pytz = "^2023.3"
pydantic = "^2.0.3"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"